+++ core/src/shape/sphere.rs
@@ -3,12 +3,12 @@ use super::{DistanceApprox, Shape};
 
 #[derive(Clone)]
 pub struct Sphere {
-    center: Point3<f64>,
-    radius: f64,
+    center: Point3<f32>,
+    radius: f32,
 }
 
 impl Sphere {
-    pub fn new(center: Point3<f64>, radius: f64) -> Self {
+    pub fn new(center: Point3<f32>, radius: f32) -> Self {
         Sphere {
             center: center,
             radius: radius,
@@ -17,11 +17,12 @@ impl Sphere {
 }
 
 impl Shape for Sphere {
-    fn contains(&self, p: Point3<f64>) -> bool {
+    // Overwrite default method for performance
+    fn contains(&self, p: Point3<f32>) -> bool {
         (self.center - p).magnitude2() <= (self.radius * self.radius)
     }
 
-    fn distance(&self, p: Point3<f64>) -> DistanceApprox {
+    fn distance(&self, p: Point3<f32>) -> DistanceApprox {
         let d = (self.center - p).magnitude() - self.radius;
         DistanceApprox {
             min: d,
